Remove limit on max size of PNOR TOC
Previously we assumed the TOC would never be larger than 4kb, this
limited us to <31 sections in pnor. We are getting to the point that is
not enough. To satify the github issue a change was needed in the ffs
tool to correctly build the TOC so it describes its size correctly.
Those changes can be found here: github.com/open-power/ffs/pull/14
The changes in this commit fix hostboot so that it will correctly
parse TOCs > 4kb.

Resolves open-power/boston-openpower/#1307

Change-Id: Id2d8c250b0b1e6b87adfd4d927119fe006e8910d
RTC: 180032
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/62643
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Disable-CI: STEWART E. SMITH <stewart@linux.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>