Add Slave-Procs-Only SBE Update via FSI I2C in istep 8.1
This change adds the ability to call the SBE Update function and
have it run only against the slave processors in the system early
in the IPL when they can only access the SBE Seeproms via FSI I2C.
This ability is called in istep 8.1.  There are some other minor
changes to support this, as well.

Change-Id: Id6b205b8b4886f1e9f8cbe25518f894bb996a6b3
CQ:SW423354
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/56798
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>