ipmi: Remove IpmiRP dependency on targeting
Shortly, a PNOR implementation will be introduced that is dependent on
IPMI for managing the layout of the LPC FW space. To facilitate this,
remove IpmiRP's dependency on Targeting so that the ipmibase module can
be moved into the base image (to be initialised before the PNOR
implementation).

Break the dependency by having the Targeting service initialisation
fetch the maximum IPMI buffer size from the IPMI interface and populate
the necessary attribute for HBRT.

Change-Id: Ie51244435b0e14c9a5f0838283021b08858035f6
Signed-off-by: Andrew Jeffery <andrewrj@au1.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65937
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67694
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>