ipmi: IpmiDD and IpmiRP must never free resources
An IPMI-based PNOR implementation will be introduced in the future. If
it is in use (and it likely will be), then IpmiDD and IpmiRP must be
functional right up until the shutdown() syscall is invoked.  This is
necessary because hostboot assumes parts of the PNOR can be paged in at
any time (though not necessarily dirtied and flushed), and IPMI is now
required to fulfil the requests.

Remove the unnecessary shutdown-related code from both classes.

Change-Id: I82276e5e12c8b917d655000a456b72a2ccacdb5f
Signed-off-by: Andrew Jeffery <andrewrj@au1.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67077
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67692
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>