ipmi: Increase polling rate to decrease boot time
The recent introduction of the IPMI-based HIOMAP transport for the PNOR
means the boot process is now heavily constrained by the speed of IPMI
transfers.

As an interim change to introducing interrupt-driven BT support,
decrease the period of the IPMI poller to increase boot speed by a
factor of about 1.7x when using the IPMI HIOMAP transport. This also
slightly increases the speed of a MBOX HIOMAP boot, but the point here
is that there's not a detrimental effect in this configuration.

Change-Id: I853ab013e33715101383c2c80c2d1b67c1c79265
Signed-off-by: Andrew Jeffery <andrewrj@au1.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68321
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>