Add missing mutex in LPC error path
There is a pre-check to look for errors before every LPC read or
write operation.  This check was running outside of the mutex,
which meant that if an operation caused an error and then another
thread attempted a separate LPC op, the second thread would end
up reporting the error from the first operation,  This would
typically just result in a double log, but there are cases where
an error is expected (the SIO problem for the hiomap protocol)
and is deleted by the caller.  Because of the second thread
seeing the error, the error condition ends up being a visible
log.

The change here is to move the pre-check inside the mutex so that
only 1 thread can ever be checking the LPC status at a time.

Change-Id: I9ce0ce48252b7d2b271aa5dd6e0a819dfb728a25
CQ: SW450825
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68716
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>