Ignore 'no error found' RC from p9_check_idle_stop_done
The p9_check_idle_stop_done HWP returns a non-zero RC for all
executions, even if it did not find an error.  This is a problem
because we will commit the generated log for the passing case,
and it shows up as visible.  Note that this behavior was
changed in later releases.

The change here is to explicitly check for the 'no error' RC
of RC_CORE_POWERED_AND_RUNNING and ignore it.

Change-Id: Idb3564385e0f6b4d6d45fdd35ac06f58e284580a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71683
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>