Again try to fix checkpoint issue with fips930 simics
Previously we tried to fix the issue where checkpoints were failing
because environment variables were not set. We thought by using the
printenv command it would just return "" if no variable was defined
but we still saw errors. Now we will explicitly set the local variable
to "" then check if the environment variable exists, if it does
then we will set it to the local variable we are using.

Change-Id: Iff5ffc9d9a05b1dd8ee85dbc8299884c26d4c892
CQ: SW458604
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72649
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/73019
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>