Preload CCS program to support EPOW on NVDIMM
This is written specifically for NVDIMM on P9 Nimbus. The current
recipe relies on the MC sequencer to enter STR in the event of EPOW
and could be delayed depending on the status of the bus. CCS will
ensure the immediate entry of the STR and the order of execution
(STR -> assert RESETn)

Change-Id: Ic117b489582aaf92810e905f43fe2e4e0408eeeb
CQ:SW458983
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75178
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75195
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>