Process Core and EQ Gard Records in MPIPL Path
When an error happens or is injected in a core during runtime and
MPIPL is performed following, the gard record does not get processed,
the core does not get deconfigured and then later fails to wake up
in istep 16.2. This change adds the processing of core and EQ
gard records in MPIPL path, so that they can be deconfigured
properly.

Change-Id: I66e0e7b26b5ee7ed5925202088a366a461ef4e79
CQ: SW459970
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75385
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>