Verify UCD update success at end of store to UCD flash
Implements manufacturer recommended verification procedure to determine if the
UCD updates to data flash worked correctly.  Verifies if the MFR_STATUS
bit indicating flash update completed is set, then checks to see if the
flash update error bit is confirmed not to be set.  On any failure, calls out
the UCD device as a predictive error.

Change-Id: Ibe0e00029dbd41291f8651f083434ce059c2c37f
RTC: 205979
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75143
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>