Fix deadlock in ECC error shutdown path
There is a scenario wherein if a PNOR ECC UE occurs when attempting to service a
page fault, the kernel will deadlock when it attempts to kill the task that
triggered the fault.  This is due to the kill routine being called while holding
the VMM spinlock and then collecting a backtrace, which also performs
a nested acquire of the same spinlock.  This fix inhibits invoking the
backtrace if the kill routine holds the VMM spinlock.

Change-Id: I75bf1f248740a08fd485379d88e146096edf65a9
CQ: SW461429
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75265
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>