Remove Redundant LPC Bus Check
Current implementation of lpcWrite contains a check for errors
on LPC bus before the attempt to writeLPC, and writeLPC is not
attempted if an error is found on the bus. However, the error
check performs bus reset, and the same check is performed in
writeLPC with the difference that the lpcWrite is still attempted
after the bus reset, and only if the write fails is the error
returned.

This commit removes the first check for LPC errors to give lpcWrite
a chance to run after the LPC bus reset due to errors.

Change-Id: I6375e6575f771e33bd3b866534eb3cd8656d38d8
CQ: SW455075
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76429
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>