Update SBE In Istep7.5 In SMF Mode
A bug was discovered on OpenPOWER systems wherein if SMF
is enabled, we would checkstop in istep10.1 when trying
to XSCOM to the slave proc. The issue was that SBE at that
time didn't know we were booting in SMF mode and so didn't
set up the XSCOM BAR for the slave proc appropriately.
The solution is to force SBE update in istep7.5 if SMF is
enabled in the MRW (ATTR_SMF_CONFIG) but SBE doesn't yet
know about that (bit 16 in scratch reg6 is not set).

Change-Id: Ibaa2d1e5b441f378528443aa84e5cd61789792dd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82764
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>