Remove unmask of MCBISTFIRQ[12]
This FIR bit was only used for a DD1.x workaround, and the PRD code to
handle it had already been removed from FW. We thought we would need
this FIR to implement the symbol mark performance workaround in
commit 70081 (CQ SW453101) but we didn't end up needing it. As a result
of unmasking this, there have been erroneous error logs appearing in
manufacturing tests.

Change-Id: I9ce1d563280b889609d2b0a6301ed82bd8540ba0
CQ: SW475386
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85007
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85066
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>