PRD: fix Explorer FSI address workaround in FIRDATA code
There was a previous workaround to stub out filling in the FSI address
for Explorer in the FIRDATA code for checkstop analysis. That workaround
would fill in the FSI address for processor chips before initializing
the structure that contains the FSI address. Therefore, the FSI address
for processors would always get initialized to 0xffffffff. The fix is to
simply put the code in the correct order.

Change-Id: Id76ec015ffec1565236481fcc14be6c1a3145d28
CQ: SW474413
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84228
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Paul Greenwood <paul.greenwood@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>