Add plug-rule callout to memory vpd lookup failure
If there is a mismatch in the dimm ranks behind a memory port
that will result in a VPD lookup error.  It would be better if
we got a list of the dimms but that isn't how the code is currently
designed.  In lieu of a perfect fix, we will add a procedure
callout that points a memory plugging issue.

Change-Id: I036f002c2145fd1debd957db8f46bc2e23d298f6
CQ: SW477820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85826
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85853
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>