Fix mask calculation for ATTR_PROC_PCIE_PHB_ACTIVE mask
ATTR_PROC_PCIE_PHB_ACTIVE is an attribute consumed by HDAT that
indicates which PHB targets under a PROC chip are functional. Hostboot
needs to update this mask whenever it deconfigures a PHB target to
communicate the new functional state to the hypervisor.

The mask calculation was incorrect because it used the wrong shift
operator. This commit fixes the issue.

Change-Id: Ie32e29029b50ace1b0b3cc337695cb688f642af0
CQ: SW490270
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98641
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Nicholas E Bofferding <bofferdn@us.ibm.com>