Anti-Rollback Support: Adding Secure Version to Signed Containers
This commit picks up the SIGNING_SECURITY_VERSION environment variable
in the hbfw/image/makefile and passes it genPnorImages.pl.  genPnorImages.pl
then includes it on calls to the signing tools and also puts the value
immediately before the HW Keys' Hash at the end of the HBBL binary.

RTC:254577
Change-Id: I8420a7cdca95e726921d7c3bcaa0b79a017686c6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102720
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>