Anti-Rollback Support: Updating SecureROM to check for Secure Version
This commit uses the HW params log field to pass a secure version
into the SecureROM's verify function.  This value is then used to
check against the container's FW secure version level.

It also updates the verifyContainer() functions to provide an option
to pass in a different secure version value than the default system
value.

Change-Id: Id110d22bdb7bbb1f86363a17fd1c9d0e7543d6a5
RTC:254577
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100805
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102747
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>