P9: Migration effort to RH8
Updated the env.bash to set up environment variables
that are specific to Red Hat 8 so P9 will compile
for RH8.

Removed compiling for ctags and cscope as they are not
used.  This was done for P10 therefore just remvoing
it for P9 to facilitate the migration effort.

Updated the call to Perl scripts handle_duplicate.pl,
handle_fapi_attr_mapping.pl and create_ekb_targattr.pl to
be done via Perl explicitly.  With Perl 5 dropping the
current working directory from @INC, calling these scripts
via Perl explicitly allows passing in an include directive
to point to include files needed which would otherwise fail
the script's execution if not found.

Change-Id: Ie65a8f889343bcc978f600cb90beb3f767a99c16
RTC: 296965
CQ: SW554107
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134133
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>