P9: Support "Disable TPM"
This commit supports the new user option to disable the TPM(s) in the
system.  If the option is selected, The TPMs will be safely disabled
and their status will be set to non-functional.

For FSP-based systems it will use the shared ATTR_DISABLE_TPM
attribute to know if this action is to be performed.  For BMC-based
systems it will use the IPMI interface to get the "Disable TPM" sensor
value from the BMC.

Change-Id: Iaa951a718e86184d2de566f6c640779753a1bdbf
STGD: 581453
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/162174
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>