Support dynamically determining if TPM Model is 65x or 75x
This commit adds support to dynamically determine if the TPM Model
in a system is either a Nuvoton 65x or a Nuvoton 75x. Only one
of these 2 TPM models are allowed in a system.  And, Nuvoton 75x
TPMs are only expected in systems with 1 TPM.

These are the main changes:

- A new writeable attribute ATTR_TPM_MODEL_DETERMINED has been added
to track when the TPM's model has been determined

- Where appropriate, existing reads of ATTR_TPM_MODEL have been
updated to look at the new attribute ATTR_TPM_MODEL_DETERMINED

- tpmPresence() is used to determine what the TPM model is.  It
first checks if the TPM model is 65x and then if it is 75x.

- tpmReadAttributes() has been updated to support returning back
either the 65x or 75x information depending on whether
tpmPresence() is trying to determine the TPM model or if the
TPM model has already been determined.

- Any calls to read ATTR_TPM_INFO have been checked to see if
they now need to go through tpmReadAttributes()

- Check if the TPM I2C device information that the FSP might have
put into HDAT for a TPM Model 65x device gets properly updated to
the TPM Model 75x device information if a TPM Model 75x was
dynamically found earlier in the IPL

STGD: 635054
Change-Id: I2174396ac6d2d44dba06cd5b5b78cc3a2293eb25
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/177744
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>