Fix TPM HDAT I2C Device Info for P9 OpenPower Systems
This commit takes the TPM HDAT I2C Device Info override logic
that was previously exclusive to FSP-based systems and applies it
to OpenPower systems.

The reason for this is that like FSP-based systems the raw I2C
Device information for the TPMs in the MRW are put into the HDAT
I2C Device structures for OpenPower systems.

However, the MRW logic is only aware of TPM Model 65x devices and
therefore the data needs to be overridden when TPM Model 75x devices
are dynamically found in the system by Hostboot.

Change-Id: Ie1ac4338d67e85df36f95221bce84818c192f8e1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/177840
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>