Always check before accessing ATTR_TPM_MODEL_DETERMINED
Due to Defect 327918 there is a possibility that on a MPIPL
following a concurrent code udpate Hostboot code might not be able
to find the new attribute ATTR_TPM_MODEL_DETERMINED.  To avoid this
causing problems, checks were added to the code to always try to
get this attribute first.  The code would then correctly handle
if the attribute was available or not available.

Change-Id: I0591e6d4ca680774647e349ec4ac327060f09df5
STGD: 646901
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/178680
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>